Move to statsmodels/stats, rename class to AnovaRM
